requirements.txt
streamlit==1.33.0
joblib==1.3.2
numpy==1.26.3
matplotlib==3.7.5
scipy==1.11.3
google-generativeai==0.4.0
python-dotenv==1.0.0